requirements.txt
python-multipart
fastapi[standard]
uvicorn
openai-whisper
